fix: keep TTS helpers out of route exports
